Implement room creation
Read user input for a Room instance and add it to the database

InputRoom reads all fields required to build a Room instance and
inserts it into the database. The input room cannot be a Warehouse.
Make EasyInput.Select return the actual object instead of the index so
in-place modifications of the list during invocation are accounted for.