feat: Implement Appointment CRUD
This commit adds all of the functionality needed for CRUD over Appointment objects, as well as the necessary infrastructure for integration into the system.

The Patient and Doctor classes, as well as their associated boilerplate throughout the codebase, are just stubs. They exist to support the necessary abstractions for the Appointment class.

Do note that the way CRUD over Appointments currently work doesn't fit the spec entirely, as we don't have login functionality yet. The patient gets selected same as any other field, instead of being predetermined by the login user (if he is a patient). This part should be trivial to implement once we get there.

Updating is also kind of clunky, since if we decide to update all of the fields instead of just one or two, we can't cover every possible valid state from the get-go, since the way things currently work, all of the in-between states between every field's update have to be consistent as well.
Imagine if we're trying to change an appointment's doctor from doctor a to doctor b, while also planning to change the appointment's schedule date and time as well. If doctor b isn't available during the appointment's time *before* it gets changed, then we won't have an opportunity to select him and change the scheduled datetime to one he is available for (the order of changes is doctor, patient, schedule datetime). We can change this if we think it's a big deal.

Changes to EasyInput:
Unfortunately I couldn't neatly seperate the work I did on EasyInput from Appointment CRUD since I made one rely on the other before either was commited, so these changes are part of this commit.

If we ever reach a point where we are trying to select from an empty list, we won't be able to exit Select's loop unless we input the cancel string.
This issue came up while I was doing some work on Appointment CRUD, however the issue has since been eliminated from that side, hence there is (as far as I know) no cases where this issue comes up (currently). If this seems superfluous, we can remove it.