feat: Implement anti-troll mechanism for patients
The UserAccount class has been expanded to keep track of timestamps of when the user performed modifications/creations on Appointment objects. If the user account belongs to a patient, we check if the patient exceeded a specified amount of available modifications and creations for the last 30 days, and if he has, we block him from the system (on the basis of being a troll).