feat: Respond to patients' appointment requests
This commit adds functionality needed for responding to patients' requests.

Secretary can approve or deny requests for updating or deleting patients' requests.
Every approved update request is then moved to a list of appointments.
If he chooses to deny update request, it's state is changed to DENIED.
Similarly for approving and denying delete requests, their state is changed accordingly and if it's approved, then the appointment is marked as deleted.