feature: Add room splitting interface
The manager can choose to split a room when its renovating

The rooms are kept directly inside the renovation object because
we don't want them in the system until the renovation is completed.
Equipment is selected for relocating in-place. There is no mechanism
implemented that actually applies these changes. They need to be
performed in the background. Also, we have to prevent further
renovations from using a room that'll get split up, but also make sure
to warn the user if he tries to renovate-split a room before some other
renovation-split is to be performed.