feat: Add examinations to MedicalRecord
MedicalRecord now also keeps track of the examinations associated with the patient.
We don't differentiate between completed and pending examinations because as far as I understand doctors can perform examinations that may not necessarily be due *right now*, so a chronological system for separating completed from pending examinations wouldn't reflect the real situation. We would need an additional member in Appointment to signify if it has been completed or not.
Also note that I'm using the terms Appointment and Examination interchangeably, as we don't have Operations in the system.