feat: Implement basic account system
UserCommandView is an abstract class that does the heavy lifting regarding
user input (invoking commands).
ManagerCommandView is one implementation of this class that enables all
functionality for the manager.
The biggest issue right now is the need to pass all the views to a command
view in order to call their command methods. The other alternative would be
for each of the command views to create their own views, but this would then
require services and other views, so it doesn't seem feasible.